(feat) Adds navigation bar to create habit and inbox views
