--- app/containers/app.js
@@ -1,75 +1,0 @@
-var React = require('react');
-var Text = require('react-native').Text;
-var View = require('react-native').View;
-var Component = require('react-native').Component;
-var StyleSheet = require('react-native').StyleSheet;
-var TouchableHighlight = require('react-native').TouchableHighlight;
-var Navigator = require('react-native').Navigator;
-
-var Welcome = require('../components/welcome');
-var AddHabit = require('./addHabitsApp');
-var Habits = require('./habitsApp');
-
-var App = React.createClass({
-  render: function () {
-    return (
-     <Navigator
-        initialRoute = {{id: 'Welcome', name: 'Index'}}
-        renderScene = {this.renderScene}
-     />
-     )
-  },
-  renderScene: function (route, navigator) {
-    var routeId = route.id;
-    if(routeId === 'Welcome') {
-      return (
-        <Welcome 
-            navigator={navigator} />
-        )
-    }
-    if(routeId === 'AddHabit') {
-      return (
-        <AddHabit 
-            navigator={navigator} />
-        )
-    }
-    if(routeId === 'Habits') {
-      return(
-        <Habits 
-            navigator={navigator} />
-        )
-    }
-  }
-});
-
-var styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#FFFFFF'
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10
-  },
-  instructions: {
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5
-  },
-  button: {
-    height: 30,
-    width: 80,
-    alignItems: 'center',
-    alignSelf: 'center',
-    borderColor: 'gray',
-    borderWidth: 1,
-    borderRadius: 5,
-    padding: 5,
-    margin: 20
-  }
-});
-
-module.exports = App;
